Changed test to reuse a test calendar in order to stay with quota - 25 creations

+++ src/test/java/org/rrabarg/teamcaptain/fixture/ScheduleFixture.java
@@ -3,6 +3,8 @@ package org.rrabarg.teamcaptain.fixture;
 import static java.time.temporal.ChronoUnit.HOURS;
 
 import java.io.IOException;
+import java.net.Inet4Address;
+import java.net.UnknownHostException;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
@@ -10,11 +12,15 @@ import org.rrabarg.teamcaptain.Match;
 import org.rrabarg.teamcaptain.MatchBuilder;
 import org.rrabarg.teamcaptain.ScheduleRepository;
 import org.rrabarg.teamcaptain.ScheduleService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
 public class ScheduleFixture {
+
+    Logger log = LoggerFactory.getLogger(this.getClass());
 
     String aTitle = "A test match";
     LocalDate aDate = LocalDate.of(2011, 3, 20);
@@ -30,13 +36,19 @@ public class ScheduleFixture {
     ScheduleRepository scheduleRepository;
     private String scheduleId;
 
-    public void createSchedule() throws IOException {
-        scheduleId = scheduleRepository.addSchedule("Test schedule "
-                + System.currentTimeMillis());
+    public void deleteSchedule() throws IOException {
+        if (scheduleId != null) {
+            scheduleRepository.deleteSchedule(scheduleId);
+        }
     }
 
-    public void deleteSchedule() throws IOException {
-        scheduleRepository.deleteSchedule(scheduleId);
+    public void clearSchedule() throws IOException {
+        scheduleRepository.clearSchedule(scheduleId);
+    }
+
+    public void clearTestSchedules() throws IOException {
+        scheduleRepository
+                .clearSchedulesWithSummaryStartingWith("Test schedule ");
     }
 
     public void scheduleMatch() {
@@ -49,11 +61,34 @@ public class ScheduleFixture {
                 .withLocation(aLocationFirstLine, aLocationPostcode).build();
     }
 
-    public void reset() throws IOException {
+    public void reset() throws IOException, InterruptedException {
+        teardown();
+        setup();
+    }
+
+    public void setup() throws IOException, InterruptedException {
+
         if (scheduleId != null) {
-            deleteSchedule();
+            return;
         }
-        createSchedule();
+
+        scheduleId = scheduleRepository.findSchedule(getTestScheduleName());
+        if (scheduleId == null) {
+            scheduleId = scheduleRepository.addSchedule(getTestScheduleName());
+            log.info("Created test schedule " + scheduleId);
+        } else {
+            log.info("Found test schedule " + scheduleId);
+        }
+    }
+
+    public void teardown() throws IOException, InterruptedException {
+        if (scheduleId != null) {
+            clearSchedule();
+        }
+    }
+
+    private String getTestScheduleName() throws UnknownHostException {
+        return "Test schedule-" + Inet4Address.getLocalHost().getHostAddress();
     }
 
 }
